examples/DeepLabv3/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
PySide6>=6.5.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
onnx>=1.16.0
onnxruntime>=1.17.0
